third_party/WebKit/LayoutTests/bluetooth/server/getPrimaryService/gen-service-not-found.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return getHealthThermometerDevice({
      filters: [{services: ['health_thermometer']}],
      optionalServices: ['glucose']})
    .then(({device}) => assert_promise_rejects_with_message(
      device.gatt.getPrimaryService('glucose'),
      new DOMException(
        'No Services matching UUID ' + glucose.uuid + ' found in Device.',
        'NotFoundError')));
}, 'Request for absent service. Reject with NotFoundError.');

</script>
